joined test bases

+++ src/test/java/tests/TestBase.java
@@ -9,41 +9,66 @@ import io.qameta.allure.selenide.AllureSelenide;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
-
+import java.util.Arrays;
 import java.util.Objects;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.AfterEachCallback;
+import org.junit.jupiter.api.extension.BeforeAllCallback;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import static com.codeborne.selenide.WebDriverRunner.getWebDriver;
+@Tag("test")
+@ExtendWith(TestBase.TestBaseExtension.class)
 
 public class TestBase {
 
     private static final LaunchConfig CONFIG = LaunchConfigReader.Instance.read();
 
     @BeforeAll
-    @Tag("ui")
-    static void setUpUI() {
+    static void setUp() {
         SelenideLogger.addListener("AllureSelenide", new AllureSelenide());
 
         ProjectConfiguration projectConfiguration = new ProjectConfiguration(CONFIG);
-        projectConfiguration.webConfig();
+        if (isUiTest()) {
+            projectConfiguration.webConfig();
+        } else if (isApiTest()) {
+            projectConfiguration.apiConfig();
+        }
     }
 
     @AfterEach
-    @Tag("ui")
-    void tearDownUI() {
-        Attach.screenshotAs("Screenshot");
-        Attach.pageSource();
-        if (Objects.equals(System.getProperty("browserName"), "chrome")) {
-            Attach.browserConsoleLogs();
-        }
-        Attach.addVideo();
+    void tearDown() {
+        if (isUiTest()) {
+            Attach.screenshotAs("Screenshot");
+            Attach.pageSource();
+            if (Objects.equals(System.getProperty("browserName"), "chrome")) {
+                Attach.browserConsoleLogs();
+            }
+            Attach.addVideo();
 
-        getWebDriver().manage().deleteAllCookies();
+            getWebDriver().manage().deleteAllCookies();
+        }
     }
 
-    @BeforeAll
-    @Tag("api")
-    static void setUpAPI() {
-        ProjectConfiguration projectConfiguration = new ProjectConfiguration(CONFIG);
-        projectConfiguration.apiConfig();
+    private static boolean isUiTest() {
+        return Arrays.stream(TestBase.class.getAnnotationsByType(Tag.class))
+                .anyMatch(tag -> tag.value().equals("ui"));
+    }
+
+    private static boolean isApiTest() {
+        return Arrays.stream(TestBase.class.getAnnotationsByType(Tag.class))
+                .anyMatch(tag -> tag.value().equals("api"));
+    }
+
+    public static class TestBaseExtension implements BeforeAllCallback, AfterEachCallback {
+        @Override
+        public void beforeAll(ExtensionContext extensionContext) throws Exception {
+            // Implement the setup logic
+        }
+
+        @Override
+        public void afterEach(ExtensionContext extensionContext) throws Exception {
+            // Implement the teardown logic
+        }
     }
 }
